code cleanup
hostname is a String, not a weird charlist in the configs

introduc the prefix. A user can connect to multiple databases or even use different settings for the same database. The prefix being a simple namespace for connections and user configurations

fix the route refresh after I broke it when I introduced the prefix namespace

docker-compose, for testing with a cluster of 3

the router test is global, not anymore meant for the routing-only implementations. Routing specific tests should be tagged indivitually with ":routing"

update the .iex, for convenience

removed the fuzzyurl lib. It would've break when the newer schema "neo4j" will become official

heh ... cleanup, sorry for the mess

fixed some dialyzer errors and refactored some low-hanging fruits

code cleanup

work in progress; https://github.com/florinpatrascu/bolt_routing/issues/1

optimization

Fixes #1 - this was a tough one and there's no going back to 1.x from this, b/c it's better =)


basic documentatino skeleton and moving stuff around; #3, WIP

minimialist support for running tests containing Python stubs

keeping FixEx around for awhile and will clean after

using porcelain ... but :'(

- added a stub script that conforms to our driver implementation - thanks Dominique for catching this.
- fixed the output stream capture when the boltstub dies or exits prematurely (or on some success conditions)

code cleanup started, will group our own stubs in a different folder, until we refactor for the pipeline. There' still value in using the stub, even for not needing docker, during the tests.

using stubs, is not that bad :)

bug fixes and improvements

move code around before cleaning

add back the default connection, for the rest of the tests


test updates

keep just the stubs we use

finally, controlling the stubs

error/exception disambiguation and lots of code cleaning

we are even closer to the official behavior, in respect of reporting troubles related to the routing
rewording the changelog

workig on the docs

configuration docs